Role is completed

+++ Modules/Core/Http/Controllers/RolesController.php
@@ -23,6 +23,14 @@ class RolesController extends Controller {
 
 	//------------------------------------------------------
 	public function index() {
+
+		if ( ! \Auth::user()->hasPermission( $this->data->permission->prefix,
+			$this->data->permission->pretext . "read" )
+		) {
+			return redirect()->route('core.backend.dashboard')
+				->withError([getConstant( 'permission.denied' )]);
+		}
+
 		$this->data->title = "Roles";
 
 		return view( $this->data->view . "index" )
@@ -31,18 +39,26 @@ class RolesController extends Controller {
 
 	//------------------------------------------------------
 	public function getList( Request $request ) {
+
+		if ( ! \Auth::user()->hasPermission( $this->data->permission->prefix,
+			$this->data->permission->pretext . "read" )
+		) {
+			$response['status']   = 'failed';
+			$response['errors'][] = getConstant( 'permission.denied' );
+
+			return response()->json( $response );
+		}
+
 		$list = Role::withCount( [ 'users', 'permissions' ] );
 		if ( $request->has( "s" ) ) {
 			$list->where( "name", "like", "%" . $request->get( 's' ) . "%" );
 			$list->withTrashed();
 		} elseif ( $request->has( "trashed" ) && $request->get( "trashed" ) == 1 ) {
 			$list->withTrashed();
-			$list->orderBy("deleted_at", "desc");
-		} else
-		{
+			$list->orderBy( "deleted_at", "desc" );
+		} else {
 			$list->orderBy( "created_at", 'desc' );
 		}
-
 		$config             = \Config::get( "core" );
 		$data['list']       = $list->paginate( $config['settings']->records_per_page );
 		$data['trashCount'] = Role::onlyTrashed()->count();
@@ -80,8 +96,8 @@ class RolesController extends Controller {
 				if ( $item->enable == 1 ) {
 					$item->enable = 0;
 				} else {
-					$item->enable = 1;
-					$item->deleted_at = NULL;
+					$item->enable     = 1;
+					$item->deleted_at = null;
 				}
 			}
 			$item->save();
@@ -96,12 +112,20 @@ class RolesController extends Controller {
 	}
 
 	//------------------------------------------------------
-	public function create() {
-		return view( 'core::create' );
-	}
 
 	//------------------------------------------------------
-	public function store( Request $request ) {
+	public function store( Request $request )
+	{
+
+		if ( ! \Auth::user()->hasPermission( $this->data->permission->prefix,
+			$this->data->permission->pretext . "update" )
+		) {
+			$response['status']   = 'failed';
+			$response['errors'][] = getConstant( 'permission.denied' );
+
+			return response()->json( $response );
+		}
+
 		$rules     = array(
 			'name' => 'required|max:50',
 		);
@@ -152,48 +176,52 @@ class RolesController extends Controller {
 	}
 
 	//------------------------------------------------------
-	public function edit() {
-		return view( 'core::edit' );
-	}
 
 	//------------------------------------------------------
-	public function update( Request $request)
-	{
-
+	public function update( Request $request ) {
 		if ( ! \Auth::user()->hasPermission( $this->data->permission->prefix,
 			$this->data->permission->pretext . "update" )
 		) {
 			return redirect()->route( $this->data->route )
 			                 ->withErrors( [ getConstant( 'permission.denied' ) ] );
 		}
-
 		$rules = array(
-		    'id' => 'required',
-		    'slug' => 'required|unique:core_roles,slug,'.$this->data->input['id'],
+			'id'   => 'required',
+			'slug' => 'required|unique:core_roles,slug,' . $this->data->input['id'],
 		);
-
-		$validator = \Validator::make( $this->data->input, $rules);
+		$validator = \Validator::make( $this->data->input, $rules );
 		if ( $validator->fails() ) {
-		        $errors             = $validator->errors();
-		        $response['status'] = 'failed';
-		        $response['errors'] = $errors;
-		        return response()->json( $response );
+			$errors             = $validator->errors();
+			$response['status'] = 'failed';
+			$response['errors'] = $errors;
+
+			return response()->json( $response );
 		}
 		$item = Role::withTrashed()->findOrFail( $this->data->input['id'] );
 		try {
-			$item->fill($this->data->input);
+			$item->fill( $this->data->input );
 			$item->save();
 			$response['status'] = 'success';
 		} catch ( Exception $e ) {
 			$response['status']   = 'failed';
 			$response['errors'][] = $e->getMessage();
 		}
-		return response()->json( $response );
 
+		return response()->json( $response );
 	}
 
 	//------------------------------------------------------
 	public function delete( Request $request, $id ) {
+
+		if ( ! \Auth::user()->hasPermission( $this->data->permission->prefix,
+			$this->data->permission->pretext . "update" )
+		) {
+			$response['status']   = 'failed';
+			$response['errors'][] = getConstant( 'permission.denied' );
+
+			return response()->json( $response );
+		}
+
 		$item = Role::withTrashed()->findOrFail( $id );
 		try {
 			$item->delete();
@@ -208,6 +236,16 @@ class RolesController extends Controller {
 
 	//------------------------------------------------------
 	public function restore( Request $request, $id ) {
+
+		if ( ! \Auth::user()->hasPermission( $this->data->permission->prefix,
+			$this->data->permission->pretext . "update" )
+		) {
+			$response['status']   = 'failed';
+			$response['errors'][] = getConstant( 'permission.denied' );
+
+			return response()->json( $response );
+		}
+
 		$item = Role::withTrashed()->findOrFail( $id );
 		try {
 			$item->restore();
@@ -222,6 +260,16 @@ class RolesController extends Controller {
 
 	//------------------------------------------------------
 	public function deletePermanent( $id ) {
+
+		if ( ! \Auth::user()->hasPermission( $this->data->permission->prefix,
+			$this->data->permission->pretext . "update" )
+		) {
+			$response['status']   = 'failed';
+			$response['errors'][] = getConstant( 'permission.denied' );
+
+			return response()->json( $response );
+		}
+
 		$item = Role::withTrashed()->findOrFail( $id );
 		try {
 			$item->permissions()->detach();
@@ -235,16 +283,25 @@ class RolesController extends Controller {
 
 		return response()->json( $response );
 	}
-	//------------------------------------------------------
-	public function readDetails($id)
-	{
-		$this->data->item = Role::withTrashed()->findOrFail( $id );
 
+	//------------------------------------------------------
+	public function readDetails( $id ) {
+
+		if ( ! \Auth::user()->hasPermission( $this->data->permission->prefix,
+			$this->data->permission->pretext . "read" )
+		) {
+			$response['status']   = 'failed';
+			$response['errors'][] = getConstant( 'permission.denied' );
+
+			return response()->json( $response );
+		}
+
+		$this->data->item = Role::withTrashed()->findOrFail( $id );
 		try {
-			$data = view( $this->data->view . "partials.item-table" )
+			$data               = view( $this->data->view . "partials.item-table" )
 				->with( "data", $this->data )->render();
 			$response['status'] = 'success';
-			$response['data'] = $data;
+			$response['data']   = $data;
 		} catch ( Exception $e ) {
 			$response['status']   = 'failed';
 			$response['errors'][] = $e->getMessage();
@@ -252,16 +309,25 @@ class RolesController extends Controller {
 
 		return response()->json( $response );
 	}
+
 	//------------------------------------------------------
-	public function stats($id)
+	public function stats( $id )
 	{
+		if ( ! \Auth::user()->hasPermission( $this->data->permission->prefix,
+			$this->data->permission->pretext . "update" )
+		) {
+			$response['status']   = 'failed';
+			$response['errors'][] = getConstant( 'permission.denied' );
+
+			return response()->json( $response );
+		}
+
 		$this->data->item = Role::withTrashed()->withCount( [ 'users', 'permissions' ] )->findOrFail( $id );
-
 		try {
-			$data = view( $this->data->view . "partials.item-stats" )
+			$data               = view( $this->data->view . "partials.item-stats" )
 				->with( "data", $this->data )->render();
 			$response['status'] = 'success';
-			$response['data'] = $data;
+			$response['data']   = $data;
 		} catch ( Exception $e ) {
 			$response['status']   = 'failed';
 			$response['errors'][] = $e->getMessage();
@@ -269,23 +335,32 @@ class RolesController extends Controller {
 
 		return response()->json( $response );
 	}
+
 	//------------------------------------------------------
-	public function permissions($id)
+	public function permissions( Request $request, $id )
 	{
-		$this->data->item = Role::withTrashed()->findOrFail( $id );
-		$this->data->item->permission_ids = $this->data->item->permissions()->getRelatedIds();
-		$this->data->permissions = Permission::disabled()->get();
 
-		$data = view( $this->data->view . "partials.item-permissions" )
-			->with( "data", $this->data )->render();
+		if ( ! \Auth::user()->hasPermission( $this->data->permission->prefix,
+			$this->data->permission->pretext . "read" )
+		) {
+			$response['status']   = 'failed';
+			$response['errors'][] = getConstant( 'permission.denied' );
 
-		return $data;
+			return response()->json( $response );
+		}
 
+		$this->data->item                 = Role::withTrashed()->findOrFail( $id );
+		$this->data->item->permission_ids = $this->data->item->permissions()->getRelatedIds()->toArray();
+		$permissions                      = Permission::enabled();
+		if ( $request->has( 's' ) ) {
+			$permissions->where( "name", "like", "%" . $request->get( 's' ) . "%" );
+		}
+		$this->data->permissions = $permissions->get();
 		try {
 			$data = view( $this->data->view . "partials.item-permissions" )
 				->with( "data", $this->data )->render();
 			$response['status'] = 'success';
-			$response['data'] = $data;
+			$response['data']   = $data;
 		} catch ( Exception $e ) {
 			$response['status']   = 'failed';
 			$response['errors'][] = $e->getMessage();
@@ -293,5 +368,37 @@ class RolesController extends Controller {
 
 		return response()->json( $response );
 	}
+
+	//------------------------------------------------------
+	public function permissionsToggle( Request $request, $id, $permission_id )
+	{
+
+		if ( ! \Auth::user()->hasPermission( $this->data->permission->prefix,
+			$this->data->permission->pretext . "update" )
+		) {
+			$response['status']   = 'failed';
+			$response['errors'][] = getConstant( 'permission.denied' );
+
+			return response()->json( $response );
+		}
+
+		$item = Role::withTrashed()->findOrFail( $id );
+		$exist = $item->permissions->contains( $permission_id );
+		try {
+			if ( $exist ) {
+				$item->permissions()->detach( $permission_id );
+			} else {
+				$item->permissions()->attach( $permission_id );
+			}
+			$response['status'] = 'success';
+		} catch ( Exception $e ) {
+			$response['status']   = 'failed';
+			$response['errors'][] = $e->getMessage();
+		}
+
+		return response()->json( $response );
+	}
+	//------------------------------------------------------
+	//------------------------------------------------------
 	//------------------------------------------------------
 }
